kind: skip tests that require PVC support
The functional tests assume that local storage has been
pre-provisioned, as indicated in the
[kubevirtci storage doc](https://github.com/kubevirt/kubevirtci/blob/master/cluster-provision/STORAGE.md#storage-providers-inside-the-generated-images).

However, this is not true for the kind provider; it does not create
the `local-volume-provisioner` - and as a result, we're missing the
`local` storage class (which is the default one on functional
tests).

Furthermore, the kind provider setup is not creating the bind mounts
refered in the doc above.

Until the kind provider setup is updated (deploying the local
volume provisioner, and creating the bind mounts), we need to skip
these failing tests.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>