masquerade/bridge binding: use pre-provisioned tap device
Use the virt-handler pod (which has full capabilities) to create
and configure a tap device on the virt-launcher pod.

With this change, libvirt will use a non-managed, pre-configured
tap device (which does not require NET_ADMIN capabilities) instead
of creating a new one (which does require NET_ADMIN capabilities).

This commit is part of the change-set to remove the NET_ADMIN
capability from the virt-launcher pods.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>