ipv6, ra, selinux: use virt-launcher context
It is not possible to read the ancillary data from the unix domain
socket whenever it is created w/ the spc_t context, since selinux
truncates the unix control messages. The file description is sent
via the unix control messages.

The following strace is found without this patch:
```
recvmsg(5, {
    msg_name={
        sa_family=AF_UNIX,
        sun_path="/proc/29084/root/var/run/kubevirt/sockets/k6t-eth0"
    },
    msg_namelen=112->90,
    msg_iov=[{iov_base="\0", iov_len=1}],
    msg_iovlen=1,
    msg_controllen=0,
    msg_flags=MSG_CTRUNC
}, 0) = 1
```

And with this PR, it can be seen that the unix control messages
can be correctly received by the virt-launcher process:
```
recvmsg(6, {
    msg_name={
        sa_family=AF_UNIX,
        sun_path="/proc/29084/root/var/run/kubevirt/sockets/k6t-eth0"
    },
    msg_namelen=112->19,
    msg_iov=[{iov_base="\0", iov_len=1}],
    msg_iovlen=1,
    msg_control=[{cmsg_len=20, cmsg_level=SOL_SOCKET, cmsg_type=SCM_RIGHTS, cmsg_data=[7]}],
    msg_controllen=24,
    msg_flags=0
}, 0) = 1
```

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>