network, bind mech: add launcher PID & queue number attributes
This commit adds the launcher PID and queue number attributes to
the respective bind mechanisms; slirp and macvtap do not use the
queueNumber attribute.

This will enable to `BindMechanism` interface to be simplified,
thus making the split of it into two different interfaces: one
executed in the trusted `virt-handler` pod, and another on the
untrusted `virt-launcher` pod.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>