network, bind mech: provide separate funcs for each phase
This commit creates new functions that will be used for the bind
mechanism phase#1, which is executed in the virt-handler pod.

It also refactors the `getPhase1Binding` function to use these
new BindMechanism factory functions.

Ultimately, in a follow up PR, these will be moved to the
virt-handler pkg, thus effectively separating the bind mechanism
in two interfaces.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>