network, bind mech: load cached interface uses struct attributes
Refactor the `loadCachedInterface` in all the bind
mechanisms to use the interface name attribute defined in the
structs that implement them.

Unfortunately, using the PID attribute from the struct requires
more work, since both stages of VM networking configuration are
using different PIDs - to access the mount namespace of the
launcher pod.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>