network, bind mech: load cached vif uses struct attrs
The PID must still be passed as an argument, to allow unit tests,
since the production code caches the VIF in one path, and reads the
cache from another (symlinked) path. On unit tests, this symlink
does not exist, and is overkill. A smarter scheme must be devised to
allow this parameter to be dropped - use an interface to export /
import the VIF and not rely on the file system on unit tests are on
the top of my option list.

Also simplify the interface method; it now replies only with an
error - the `exists` boolean it returned made no sense.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>